CI: Use the composer-installed version of phpunit to run tests
When phpunit was running on the PHP 7 Travis CI worker, it reported
"No tests executed!". There may be a compatibility issue with
our test setup and the default version of phpunit provided by the
PHP 7 Travis CI worker.

This commit updates the test execution command to use the
local composer-installed version of phpunit to run the tests.